refactor: 네비게이션 관련 모델 클래스 패키지 위치 변경 (NavigationItem, NavigationIconModel)
